Make core component W3C compliant
This rewrites old border,align,valign,... attributes to make the W3C XHTML parser happy. Since align is obsolete, it has been replaced with text-align.

This does not change the default layouts that can be accessed from the backend of Joomla, but it is possible to create HTML5-compliant designs without altering the source files now.